Refactor memory handling and qualifiers:
- Simplify `resize` calls in `CodeGen.Constant.cpp` and `CodeGen.Cast.cpp`.
- Use `const` qualifiers for improved code clarity in `JIT.cpp` and `CodeGenFunctionHelpers.cpp`.
- Update function declarations in `TypeGen.h` and `TypeGen.Process.cpp` for consistency.

+++ src/Compiler/CodeGen/Ast/CodeGen.Cast.cpp
@@ -297,7 +297,7 @@ namespace
 
             SWC_ASSERT(dstSize <= std::numeric_limits<uint32_t>::max());
             SmallVector<std::byte> typedNullBytes;
-            typedNullBytes.resize(static_cast<size_t>(dstSize));
+            typedNullBytes.resize(dstSize);
             std::memset(typedNullBytes.data(), 0, typedNullBytes.size());
 
             const SemaNodeView srcConstView = codeGen.viewTypeConstant(srcNodeRef);
